+++ owiddata/owiddata-stats.json
@@ -1,4 +1,5 @@
 {
+<<<<<<< Updated upstream
   "owid_RNA_count": 3,
   "owid_RNA_countries": 167,
   "owid_RNA_map": "https://github.com/greenelab/covid19-review/raw/5eeb401b6b2cf424cfaf47e86708f9c873bd6242/owiddata/maps/RNA.png",
@@ -30,4 +31,12 @@
   "viper_approved_subunit": "| Vaccine                                   | Company                                                    | Platform        |\n|:------------------------------------------|:-----------------------------------------------------------|:----------------|\n| Zifivax                                   | Anhui Zhifei Longcom                                       | protein subunit |\n| Noora vaccine                             | Bagheiat-allah University of Medical Sciences              | protein subunit |\n| Corbevax                                  | Biological E Limited                                       | protein subunit |\n| Abdala                                    | Center for Genetic Engineering and Biotechnology (CIGB)    | protein subunit |\n| Soberana 02                               | Instituto Finlay de Vacunas Cuba                           | protein subunit |\n| Soberana Plus                             | Instituto Finlay de Vacunas Cuba                           | protein subunit |\n| Covifenz                                  | Medicago                                                   | VLP             |\n| MVC-COV1901                               | Medigen                                                    | protein subunit |\n| Recombinant SARS-CoV-2 Vaccine (CHO Cell) | National Vaccine and Serum Institute                       | protein subunit |\n| Nuvaxovid                                 | Novavax                                                    | protein subunit |\n| Razi Cov Pars                             | Razi Vaccine and Serum Research Institute                  | protein subunit |\n| COVOVAX (Novavax formulation)             | Serum Institute of India                                   | protein subunit |\n| TAK-019 (Novavax formulation)             | Takeda                                                     | protein subunit |\n| SpikoGen                                  | Vaxine/CinnaGen Co.                                        | protein subunit |\n| Aurora-CoV                                | Vector State Research Center of Virology and Biotechnology | protein subunit |\n| EpiVacCorona                              | Vector State Research Center of Virology and Biotechnology | protein subunit |",
   "viper_approved_whole_virus": "| Vaccine                  | Company                                                   |\n|:-------------------------|:----------------------------------------------------------|\n| Covaxin                  | Bharat Biotech                                            |\n| KoviVac                  | Chumakov Center                                           |\n| Turkovac                 | Health Institutes of Turkey                               |\n| FAKHRAVAC (MIVAC)        | Organization of Defensive Innovation and Research         |\n| QazVac                   | Research Institute for Biological Safety Problems (RIBSP) |\n| KCONVAC                  | Shenzhen Kangtai Biological Products Co                   |\n| COVIran Barekat          | Shifa Pharmed Industrial Co                               |\n| Covilo                   | Sinopharm (Beijing)                                       |\n| Inactivated (Vero Cells) | Sinopharm (Wuhan)                                         |\n| CoronaVac                | Sinovac                                                   |\n| VLA2001                  | Valneva                                                   |",
   "viper_vaccine_counts": "38"
+=======
+  "owid_commit_date": "2022-05-16T10:10:28Z",
+  "owid_commit_date_pretty": "May 16, 2022",
+  "owid_commit_sha": "34ed8da4ae9f27f816334e64b3a3704e9b50b2a4",
+  "owid_most_recent_date": "May 15, 2022",
+  "owid_total_vaccinations": "12 billion",
+  "owid_vaccine_counts": "27"
+>>>>>>> Stashed changes
 }
